+++ NEWS.md
@@ -1,3 +1,7 @@
+# hintr 0.0.4
+
+* Rename validate/input endpoint to validate/baseline-individual and rename validate/baseline to validate/baseline-combined for clarity
+
 # hintr 0.0.3
 
 * Validate survey and programme data using a different endpoint from the validation of baseline data